Adding new test videos
Added leading zeros in filenames for future proofing when we might have
large numbers of videos.

Added animation to the videos to clue you in to whether the video is
playing properly or not.

Added audio noise to videos with pan effect to help test audio hardware
when setting up a new player.